[REF] account_analytic_wip: merge analytic_activity_based_cost feature to simplify dependencies
