+++ account_analytic_wip/readme/CONTRIBUTORS.rst
@@ -1,3 +1,4 @@
 * `Open Source Integrators <https://opensourceintegrators.com>`:
 
   * Daniel Reis <dreis@opensourceintegrators.com>
+  * Chandresh Thakkar <cthakkar@opensourceintegrators.com>
