--- analytic_activity_based_cost/i18n/analytic_activity_based_cost.pot
@@ -1,185 +1,0 @@
-# Translation of Odoo Server.
-# This file contains the translation of the following modules:
-# 	* analytic_activity_based_cost
-#
-msgid ""
-msgstr ""
-"Project-Id-Version: Odoo Server 14.0\n"
-"Report-Msgid-Bugs-To: \n"
-"Last-Translator: \n"
-"Language-Team: \n"
-"MIME-Version: 1.0\n"
-"Content-Type: text/plain; charset=UTF-8\n"
-"Content-Transfer-Encoding: \n"
-"Plural-Forms: \n"
-
-#. module: analytic_activity_based_cost
-#: model:ir.model.fields,field_description:analytic_activity_based_cost.field_activity_cost_rule__active
-msgid "Active"
-msgstr ""
-
-#. module: analytic_activity_based_cost
-#: model:ir.actions.act_window,name:analytic_activity_based_cost.action_activity_cost_rule
-#: model:ir.ui.menu,name:analytic_activity_based_cost.menu_activity_cost_rule
-msgid "Activity Based Cost Rules"
-msgstr ""
-
-#. module: analytic_activity_based_cost
-#: model:ir.model,name:analytic_activity_based_cost.model_activity_cost_rule
-msgid "Activity Cost Rule"
-msgstr ""
-
-#. module: analytic_activity_based_cost
-#: model:ir.model.fields,field_description:analytic_activity_based_cost.field_product_product__activity_cost_ids
-#: model_terms:ir.ui.view,arch_db:analytic_activity_based_cost.view_account_analytic_line_form_ext
-#: model_terms:ir.ui.view,arch_db:analytic_activity_based_cost.view_product_product_form
-msgid "Activity Costs"
-msgstr ""
-
-#. module: analytic_activity_based_cost
-#: model:ir.model.fields,field_description:analytic_activity_based_cost.field_activity_cost_rule__parent_id
-msgid "Activity Product"
-msgstr ""
-
-#. module: analytic_activity_based_cost
-#: model:ir.model,name:analytic_activity_based_cost.model_account_analytic_line
-msgid "Analytic Line"
-msgstr ""
-
-#. module: analytic_activity_based_cost
-#: code:addons/analytic_activity_based_cost/models/product_template.py:0
-#, python-format
-msgid "Can't have Activity Costs set if it is not a Cost Type."
-msgstr ""
-
-#. module: analytic_activity_based_cost
-#: model_terms:ir.ui.view,arch_db:analytic_activity_based_cost.view_form_activity_cost_rule
-msgid "Conditions"
-msgstr ""
-
-#. module: analytic_activity_based_cost
-#: model:ir.model.fields,field_description:analytic_activity_based_cost.field_activity_cost_rule__standard_price
-msgid "Cost"
-msgstr ""
-
-#. module: analytic_activity_based_cost
-#: model_terms:ir.ui.view,arch_db:analytic_activity_based_cost.view_form_activity_cost_rule
-msgid "Cost Generated"
-msgstr ""
-
-#. module: analytic_activity_based_cost
-#: model:ir.model.fields,field_description:analytic_activity_based_cost.field_account_analytic_line__activity_cost_id
-msgid "Cost Rule Applied"
-msgstr ""
-
-#. module: analytic_activity_based_cost
-#: model:ir.model.fields,field_description:analytic_activity_based_cost.field_activity_cost_rule__product_id
-msgid "Cost Type Product"
-msgstr ""
-
-#. module: analytic_activity_based_cost
-#: model:ir.actions.act_window,name:analytic_activity_based_cost.product_product_action_cost_type
-#: model:ir.ui.menu,name:analytic_activity_based_cost.product_product_menu_cost_type
-msgid "Cost Types"
-msgstr ""
-
-#. module: analytic_activity_based_cost
-#: model_terms:ir.actions.act_window,help:analytic_activity_based_cost.product_product_action_cost_type
-msgid "Create a new cost type product"
-msgstr ""
-
-#. module: analytic_activity_based_cost
-#: model:ir.model.fields,field_description:analytic_activity_based_cost.field_activity_cost_rule__create_uid
-msgid "Created by"
-msgstr ""
-
-#. module: analytic_activity_based_cost
-#: model:ir.model.fields,field_description:analytic_activity_based_cost.field_activity_cost_rule__create_date
-msgid "Created on"
-msgstr ""
-
-#. module: analytic_activity_based_cost
-#: model:ir.model.fields,field_description:analytic_activity_based_cost.field_activity_cost_rule__name
-msgid "Description"
-msgstr ""
-
-#. module: analytic_activity_based_cost
-#: model:ir.model.fields,field_description:analytic_activity_based_cost.field_account_analytic_line__display_name
-#: model:ir.model.fields,field_description:analytic_activity_based_cost.field_activity_cost_rule__display_name
-#: model:ir.model.fields,field_description:analytic_activity_based_cost.field_product_product__display_name
-#: model:ir.model.fields,field_description:analytic_activity_based_cost.field_product_template__display_name
-msgid "Display Name"
-msgstr ""
-
-#. module: analytic_activity_based_cost
-#: model:ir.model.fields,field_description:analytic_activity_based_cost.field_account_analytic_line__id
-#: model:ir.model.fields,field_description:analytic_activity_based_cost.field_activity_cost_rule__id
-#: model:ir.model.fields,field_description:analytic_activity_based_cost.field_product_product__id
-#: model:ir.model.fields,field_description:analytic_activity_based_cost.field_product_template__id
-msgid "ID"
-msgstr ""
-
-#. module: analytic_activity_based_cost
-#: model:ir.model.fields,help:analytic_activity_based_cost.field_activity_cost_rule__standard_price
-msgid ""
-"In Standard Price & AVCO: value of the product (automatically computed in AVCO).\n"
-"        In FIFO: value of the last unit that left the stock (automatically computed).\n"
-"        Used to value the product when the purchase cost is not known (e.g. inventory adjustment).\n"
-"        Used to compute margins on sale orders."
-msgstr ""
-
-#. module: analytic_activity_based_cost
-#: model:ir.model.fields,field_description:analytic_activity_based_cost.field_product_product__is_cost_type
-#: model:ir.model.fields,field_description:analytic_activity_based_cost.field_product_template__is_cost_type
-#: model_terms:ir.ui.view,arch_db:analytic_activity_based_cost.product_search_form_view
-msgid "Is Cost Type"
-msgstr ""
-
-#. module: analytic_activity_based_cost
-#: model:ir.model.fields,field_description:analytic_activity_based_cost.field_account_analytic_line____last_update
-#: model:ir.model.fields,field_description:analytic_activity_based_cost.field_activity_cost_rule____last_update
-#: model:ir.model.fields,field_description:analytic_activity_based_cost.field_product_product____last_update
-#: model:ir.model.fields,field_description:analytic_activity_based_cost.field_product_template____last_update
-msgid "Last Modified on"
-msgstr ""
-
-#. module: analytic_activity_based_cost
-#: model:ir.model.fields,field_description:analytic_activity_based_cost.field_activity_cost_rule__write_uid
-msgid "Last Updated by"
-msgstr ""
-
-#. module: analytic_activity_based_cost
-#: model:ir.model.fields,field_description:analytic_activity_based_cost.field_activity_cost_rule__write_date
-msgid "Last Updated on"
-msgstr ""
-
-#. module: analytic_activity_based_cost
-#: model:ir.model.fields,field_description:analytic_activity_based_cost.field_account_analytic_line__parent_id
-msgid "Parent Analytic Item"
-msgstr ""
-
-#. module: analytic_activity_based_cost
-#: model:ir.model,name:analytic_activity_based_cost.model_product_product
-msgid "Product"
-msgstr ""
-
-#. module: analytic_activity_based_cost
-#: model:ir.model,name:analytic_activity_based_cost.model_product_template
-msgid "Product Template"
-msgstr ""
-
-#. module: analytic_activity_based_cost
-#: model:ir.model.fields,field_description:analytic_activity_based_cost.field_activity_cost_rule__factor
-msgid "Qty. Factor"
-msgstr ""
-
-#. module: analytic_activity_based_cost
-#: model:ir.model.fields,field_description:analytic_activity_based_cost.field_account_analytic_line__child_ids
-msgid "Related Analytic Items"
-msgstr ""
-
-#. module: analytic_activity_based_cost
-#: model:ir.model.fields,help:analytic_activity_based_cost.field_product_product__activity_cost_ids
-msgid ""
-"This product will also generate analytic items for these Activity Costs"
-msgstr ""
