--- analytic_activity_based_cost/models/__init__.py
@@ -1,6 +1,0 @@
-# Copyright (C) 2020 Open Source Integrators
-# License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
-
-from . import activity_cost_rule
-from . import analytic_account
-from . import product_template
