--- analytic_activity_based_cost/readme/CONFIGURATION.rst
@@ -1,10 +1,0 @@
-Enable the Analytic Accounting setting.
-
-Create the Products representing the Cost Types to use:
-
-* Go to Invoicing/Accounting > ... > Products, and create Products representing the Cost Types to use.
-  On a Cost Type product, set:
-
-  * "Is Cost Type" flag: checked. This will make the "Activity Costs" tab visible
-  * In the "Activity Costs" tab, add a line for each cost line to generate, such as overhead, etc.
-    Set the standard cost to use for each unit used. Thesw will roll up to the parent cost type standard cost.
