+++ analytic_activity_based_cost/readme/DESCRIPTION.rst
@@ -1,9 +1,4 @@
 Generate Activity Based Costs (ABC), using Analytic Items.
 
-Activity Based Costing is a well documented technique to assign indirect costs to the the activities identified as cost drivers.
-This feature models cost driver usage as Analytic Items.
-When an Analytic Item is created, it may then generate additional Analytic Items for the corresponding indirect costs.
-For example, each timesheet hour logged could generate a quantity and amount of overhead assigned to that activity.
-
-This feature does not generate Account Moves.
-That should be implemented by a separate feature.
+This module is deprecated.
+The features are still available in the ``account_analytic_wip`` module.
