--- analytic_activity_based_cost/security/ir.model.access.csv
@@ -1,3 +1,0 @@
-id,name,model_id:id,group_id:id,perm_read,perm_write,perm_create,perm_unlink
-access_activity_cost_rule_user,access_activity_cost_rule_user,model_activity_cost_rule,base.group_user,1,0,0,0
-access_activity_cost_rule_mngr,access_activity_cost_rule_mngr,model_activity_cost_rule,account.group_account_manager,1,1,1,1
