--- analytic_activity_based_cost/views/analytic_account_line.xml
@@ -1,36 +1,0 @@
-<odoo>
-    <record id="view_account_analytic_line_form_ext" model="ir.ui.view">
-        <field name="name">Analytic Item: add activity cost fields</field>
-        <field name="model">account.analytic.line</field>
-        <field name="inherit_id" ref="analytic.view_account_analytic_line_form" />
-        <field name="arch" type="xml">
-
-            <group name="amount" position="after">
-                <group
-                    name="activity_cost"
-                    string="Activity Costs"
-                    groups="base.group_no_one"
-                >
-                    <field
-                        name="parent_id"
-                        readonly="True"
-                        attrs="{'invisible': ('parent_id', '=', False)}"
-                    />
-                    <field
-                        name="child_ids"
-                        readonly="True"
-                        attrs="{'invisible': ('child_ids', '=', False)}"
-                    >
-                        <tree name="child_ids_tree">
-                            <field name="name" />
-                            <field name="unit_amount" />
-                            <field name="amount" />
-                        </tree>
-                    </field>
-                </group>
-            </group>
-
-        </field>
-    </record>
-
-</odoo>
